requirements.txt
openai
python-dotenv
tqdm
ipywidgets
tiktoken
pypandoc
openai-agents
streamlit
markdown
pandarallel
openpyxl
weave
attrs
pypandoc
weave[all]
tavily-python
exa-py
plotly
pandas
numpy
seaborn
matplotlib
streamlit